concert case study

+++ demos/concert-tickets/src/main/java/de/tuda/stg/consys/demo/concert/schema/BuyTicket.java
@@ -19,7 +19,7 @@ public class BuyTicket implements Serializable {
 	}
 
 	public Ticket buyTicket() {
-		if (hall.ref().maxAudience > soldTickets.ref().value) {
+		if ((int) hall.ref().maxAudience > (int) soldTickets.ref().value) {
 			soldTickets.ref().inc();
 			return new Ticket();
 		} else {
